Rearrange classes and extend summaries
