+++ source/GreenEnergyHub.Charges.Libraries/source/GreenEnergyHub.DataHub.Charges.Clients/DefaultChargeLink/Contracts/CreateDefaultChargeLinksSucceeded.proto
