+++ source/GreenEnergyHub.Charges.Libraries/source/GreenEnergyHub.DataHub.Charges.Clients/DefaultChargeLink/IDefaultChargeLinkRequestClient.cs
@@ -1,0 +1,33 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Threading.Tasks;
+using GreenEnergyHub.DataHub.Charges.Libraries.Models;
+
+namespace GreenEnergyHub.DataHub.Charges.Libraries.DefaultChargeLink
+{
+    public interface IDefaultChargeLinkRequestClient
+    {
+        /// <summary>
+        /// Request the Charges domain to create default charge links
+        /// based on the supplied meteringPointIds entity's MeteringPointType.
+        /// </summary>
+        /// <param name="createDefaultChargeLinksDto">
+        /// Contains data needed by the Charges Domain to create default charges links.</param>
+        /// <param name="correlationId">CorrelationId specifies message context.</param>
+        Task CreateDefaultChargeLinksRequestAsync(
+            CreateDefaultChargeLinksDto createDefaultChargeLinksDto,
+            string correlationId);
+    }
+}
